告警处理

+++ frameworks/core/components_ng/syntax/lazy_for_each_node.cpp
@@ -221,7 +221,7 @@ void LazyForEachNode::MarkNeedSyncRenderTree(bool needRebuild)
 
 RefPtr<UINode> LazyForEachNode::GetFrameChildByIndex(uint32_t index)
 {
-    if (index < FrameCount()) {
+    if (index < static_cast<uint32_t>(FrameCount())) {
         auto child = builder_->CreateChildByIndex(index);
         if (child.second) {
             child.second->SetJSViewActive(true);
